+++ src/parser/inline/extension/index.test.ts
@@ -23,18 +23,18 @@ describe('Unit: parser/inline/extension/index', () => {
     });
 
     it('basic', () => {
-      assert.deepStrictEqual(inspect(parser('[#a]')), [['<a href="#index:a" rel="noopener">a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#a ]')), [['<a href="#index:a" rel="noopener">a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#a b]')), [['<a href="#index:a-b" rel="noopener">a b</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#a\\ ]')), [['<a href="#index:a" rel="noopener">a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#a\\ b]')), [['<a href="#index:a-b" rel="noopener">a b</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#[]]')), [['<a href="#index:[]" rel="noopener">[]</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#\\]]')), [['<a href="#index:]" rel="noopener">]</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#\\\\]')), [['<a href="#index:\\" rel="noopener">\\</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#*a*]')), [['<a href="#index:a" rel="noopener"><em>a</em></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#`a`]')), [['<a href="#index:`a`" rel="noopener"><code data-src="`a`">a</code></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#${a}$]')), [['<a href="#index:${a}$" rel="noopener"><span class="math notranslate" data-src="${a}$">${a}$</span></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#A]')), [['<a href="#index:A" rel="noopener">A</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a]')), [['<a href="#index:a" rel="noopener" class="index">a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a ]')), [['<a href="#index:a" rel="noopener" class="index">a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a b]')), [['<a href="#index:a-b" rel="noopener" class="index">a b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a\\ ]')), [['<a href="#index:a" rel="noopener" class="index">a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a\\ b]')), [['<a href="#index:a-b" rel="noopener" class="index">a b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#[]]')), [['<a href="#index:[]" rel="noopener" class="index">[]</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#\\]]')), [['<a href="#index:]" rel="noopener" class="index">]</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#\\\\]')), [['<a href="#index:\\" rel="noopener" class="index">\\</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#*a*]')), [['<a href="#index:a" rel="noopener" class="index"><em>a</em></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#`a`]')), [['<a href="#index:`a`" rel="noopener" class="index"><code data-src="`a`">a</code></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#${a}$]')), [['<a href="#index:${a}$" rel="noopener" class="index"><span class="math notranslate" data-src="${a}$">${a}$</span></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#A]')), [['<a href="#index:A" rel="noopener" class="index">A</a>'], '']);
     });
 
   });